Move registryuser form to Util
