+++ src/client/components/RegistryUserListPage/EditRegistryUser.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import _ from 'lodash';
 import { browserHistory } from 'react-router';
-import { RegistryUserForm } from './RegistryUserForm';
+import { RegistryUserForm } from '../Util/RegistryUserForm';
 
 import * as validateRegistryUser from '../../validation/registryUser';
